requirements.txt
accelerate==1.4.0
bitsandbytes==0.45.3
chardet==5.2.0
coverage==7.4.3
coverage==7.4.3
datasets==3.3.2
deepdiff==7.0.1
deepspeed==0.16.4
dvc[s3]==3.48.3
jupyterlab==4.3.1
langchain-community==0.3.21
langchain-core==0.3.51
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
lightning_fabric==2.3.1
loguru==0.7.3
mechanicalsoup==1.3.0
mlflow==2.11.3
mpire==2.10.0
mypy==1.8.0
nbdev==2.4.2
openpyxl==3.1.2
pandas==2.2.3
peft==0.14.0
polars==1.23.0
pre-commit==3.7.1
psycopg2-binary==2.9.9
pyarrow==19.0.0
pydantic>=2.7.4,<3.0.0
pymongo==4.3.3
pytest==8.0.2
python-dotenv==1.0.1
python-snappy==0.7.3
PyYAML==6.0.1
random-user-agent==1.0.1
requests==2.32.3
retry==0.9.2
rich==13.7.0
pytz==2024.1
requests>=2.31.0
seaborn==0.13.2
sentence-transformers==3.0.0
streamlit==1.40.2
tenacity==8.2.3
tensorboard==2.16.2
tiktoken==0.7.0
torch==2.5.1
torch_geometric==2.5.3
torchmetrics==1.6.1
transformers==4.49.0
trl==0.15.2
typer==0.9.0
vllm==0.7.3
wandb==0.19.0
weaviate-client==4.8.1
xlsxwriter==3.2.0
xmltodict==0.13.0
